refactor activity form
- refactor createactivityform into reusable section components
- keep section specific layout styling consistent
- reusable for future forms
- extract util functions to form-utils
