+++ calendar-ui/src/components/ActivityFormSections/ActivityScheduleSection.tsx
@@ -1,0 +1,164 @@
+import { UseFormReturn } from 'react-hook-form';
+import {
+  FormField,
+  FormItem,
+  FormLabel,
+  FormControl,
+  FormMessage,
+} from '../ui/form';
+import { Input } from '../ui/input';
+import { Textarea } from '../ui/textarea';
+import { Checkbox } from '../ui/checkbox';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '../ui/select';
+import { mockSchedulingStatuses } from '../../data/mockLookups';
+import type { CreateActivityRequest } from '@corpcal/shared/schemas';
+import { ActivityFormSection } from './ActivityFormSection';
+
+type FormData = CreateActivityRequest;
+
+type ActivityScheduleSectionProps = {
+  form: UseFormReturn<FormData>;
+};
+
+export const ActivityScheduleSection: React.FC<
+  ActivityScheduleSectionProps
+> = ({ form }) => {
+  return (
+    <ActivityFormSection title="Schedule">
+      <FormField
+        control={form.control}
+        name="schedulingStatusId"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>Scheduling Status</FormLabel>
+            <Select
+              onValueChange={(value) => field.onChange(parseInt(value))}
+              value={field.value?.toString()}
+            >
+              <FormControl>
+                <SelectTrigger>
+                  <SelectValue placeholder="Select scheduling status" />
+                </SelectTrigger>
+              </FormControl>
+              <SelectContent>
+                {mockSchedulingStatuses.map((status) => (
+                  <SelectItem key={status.id} value={status.id.toString()}>
+                    {status.displayName || status.name}
+                  </SelectItem>
+                ))}
+              </SelectContent>
+            </Select>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+
+      <FormField
+        control={form.control}
+        name="isAllDay"
+        render={({ field }) => (
+          <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+            <FormControl>
+              <Checkbox
+                checked={field.value}
+                onCheckedChange={field.onChange}
+              />
+            </FormControl>
+            <div className="space-y-1 leading-none">
+              <FormLabel>All Day Event</FormLabel>
+            </div>
+          </FormItem>
+        )}
+      />
+
+      <div className="grid grid-cols-2 gap-4">
+        <FormField
+          control={form.control}
+          name="startDate"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Start Date</FormLabel>
+              <FormControl>
+                <Input type="date" {...field} value={field.value || ''} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        {!form.watch('isAllDay') && (
+          <FormField
+            control={form.control}
+            name="startTime"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Start Time</FormLabel>
+                <FormControl>
+                  <Input type="time" {...field} value={field.value || ''} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        )}
+      </div>
+
+      <div className="grid grid-cols-2 gap-4">
+        <FormField
+          control={form.control}
+          name="endDate"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>End Date</FormLabel>
+              <FormControl>
+                <Input type="date" {...field} value={field.value || ''} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        {!form.watch('isAllDay') && (
+          <FormField
+            control={form.control}
+            name="endTime"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>End Time</FormLabel>
+                <FormControl>
+                  <Input type="time" {...field} value={field.value || ''} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        )}
+      </div>
+
+      <FormField
+        control={form.control}
+        name="schedulingConsiderations"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>Scheduling Considerations</FormLabel>
+            <FormControl>
+              <Textarea
+                placeholder="Enter scheduling considerations"
+                rows={4}
+                {...field}
+                value={field.value || ''}
+              />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+    </ActivityFormSection>
+  );
+};
